backend/requirements.txt
flask>=2.0.0
flask-limiter>=2.0.0
python-dotenv>=0.19.0
openai>=0.27.0
psycopg2-binary>=2.9.0
python-json-logger>=2.0.0
Flask-Cors>=4.0.1
gunicorn>=20.1.0
requests>=2.28.0
tabulate>=0.9.0
colorama>=0.4.5
jsonschema>=4.17.3
PyYAML>=6.0.2
sqlalchemy>=2.0.0
asyncpg>=0.27.0
redis>=4.5.0
aiohttp>=3.8.0
fastapi>=0.100.0
uvicorn>=0.22.0
pydantic>=2.0.0
starlette>=0.27.0
typing-extensions>=4.5.0